ai-project/requirements.txt
ultralytics
opencv-python
PyYAML
selenium
webdriver-manager
requests